v4l2: save back printf string for ioctl command

+++ v4l2/v4l2_codecbase.h
@@ -175,7 +175,7 @@ class V4l2CodecBase {
 
 #ifdef __ENABLE_DEBUG__
   protected:
-    const char* IoctlCommandString(long unsigned int command);
+    const char* IoctlCommandString(int command);
 
     uint32_t m_frameCount[2];
 #endif
